+++ freezeyt/filesaver.py
@@ -5,7 +5,7 @@ import asyncio
 
 from . import compat
 from .saver import Saver
-from .types import AbsoluteURL
+from .absolute_url import AbsoluteURL
 
 from typing import Callable, Iterable, BinaryIO