+++ freezeyt/freezer.py
@@ -6,9 +6,9 @@ from typing import Tuple, List, TypeVar, Any
 import asyncio
 import inspect
 import re
-import urllib.parse
 import pickle
 import base64
+import os
 
 from werkzeug.datastructures import Headers
 from werkzeug.http import parse_options_header, parse_list_header
@@ -19,10 +19,11 @@ import freezeyt.actions
 from freezeyt.encoding import decode_input_path, encode_file_path
 from freezeyt.filesaver import FileSaver
 from freezeyt.dictsaver import DictSaver
-from freezeyt.util import parse_absolute_url, is_external, urljoin
 from freezeyt.util import import_variable_from_module
 from freezeyt.util import InfiniteRedirection, ExternalURLError
-from freezeyt.util import UnexpectedStatus, MultiError, AbsoluteURL, TaskStatus
+from freezeyt.util import UnexpectedStatus, MultiError, TaskStatus
+from freezeyt.absolute_url import AbsoluteURL
+from freezeyt.compat import warnings_warn
 from freezeyt.compat import WSGIApplication
 from freezeyt import hooks
 from freezeyt.saver import Saver
@@ -114,7 +115,7 @@ def get_path_from_url(
 
     Both arguments should be results of parse_absolute_url.
     """
-    if is_external(url, prefix):
+    if url.is_external_to(prefix):
         raise ValueError(f'external url {url}')
 
     path = url.path
@@ -125,15 +126,13 @@ def get_path_from_url(
     result = PurePosixPath(url_to_path(path))
 
     if result.is_absolute():
-        url_text = urllib.parse.urlunsplit(url)
         raise ValueError(
-            f"Path may not be absolute: {result}(from {url_text})"
+            f"Path may not be absolute: {result}(from {url})"
         )
     assert '.' not in result.parts
     if '..' in result.parts:
-        url_text = urllib.parse.urlunsplit(url)
         raise ValueError(
-            f"Path may not contain /../ segment: {result}(from {url_text})"
+            f"Path may not contain /../ segment: {result}(from {url})"
         )
 
     return result
@@ -148,13 +147,17 @@ class Task:
     redirects_to: "Optional[Task]" = None
     reasons: set = dataclasses.field(default_factory=set)
     asyncio_task: "Optional[asyncio.Task]" = None
+    urls_redirecting_to_self: set = dataclasses.field(default_factory=set)
+    exception: Optional[Exception] = None
 
     def __repr__(self) -> str:
         return f"<Task for {self.path}, {self.status.name}>"
 
     def get_a_url(self) -> AbsoluteURL:
         """Get an arbitrary one of the task's URLs."""
-        return next(iter(self.urls))
+        # we need to ensure that get_a_url() will get the right one
+        # when there are urls redirection to itself
+        return next(iter(self.urls - self.urls_redirecting_to_self))
 
     @property
     def status(self) -> TaskStatus:
@@ -171,8 +174,22 @@ class Task:
         assert self.path not in new_collection
         new_collection[self.path] = self
 
+    def fail(self, exception):
+        if self.freezer.fail_fast:
+            raise exception
+        if self.exception is not None:
+            # The task already failed; we shouldn't do any more operations
+            # on it. If we do and they fail, raise that error directly.
+            raise exception
+        self.exception = exception
+        self.update_status(self.status, TaskStatus.FAILED)
+        self.freezer.call_hook('page_failed', hooks.TaskInfo(self))
+
 class IsARedirect(BaseException):
     """Raised when a page redirects and freezing it should be postponed"""
+
+class RedirectToSamePath(BaseException):
+    """Raised when a page redirects to url with same freezing path on disk as the page"""
 
 class IgnorePage(BaseException):
     """Raised when freezing a page should be ignored"""
@@ -304,7 +321,7 @@ class Freezer:
 
         # Decode path in the prefix URL.
         # Save the parsed version of prefix as self.prefix
-        prefix_parsed = parse_absolute_url(prefix)
+        prefix_parsed = AbsoluteURL(prefix)
         decoded_path = decode_input_path(prefix_parsed.path)
         if not decoded_path.endswith('/'):
             raise ValueError('prefix must end with /')
@@ -433,7 +450,7 @@ class Freezer:
         external_ok: bool = False,
         reason: Optional[str] = None,
     ) -> Optional[Task]:
-        if is_external(url, self.prefix):
+        if url.is_external_to(self.prefix):
             if external_ok:
                 return None
             raise ExternalURLError(f'Unexpected external URL: {url}')
@@ -466,7 +483,7 @@ class Freezer:
                 assert not url_part.startswith('/')
                 url_part = self.prefix.path + url_part
                 self.add_task(
-                    urljoin(self.prefix, url_part),
+                    self.prefix.join(url_part),
                     reason="from extra_files",
                 )
             elif kind == 'path':
@@ -479,7 +496,7 @@ class Freezer:
                     assert not url_part.startswith('/')
                     part = self.prefix.path + part
                     self.add_task(
-                        urljoin(self.prefix, part),
+                        self.prefix.join(part),
                         reason="from extra_files",
                     )
             else:
@@ -494,9 +511,37 @@ class Freezer:
         status_action = task.response_headers.get('Freezeyt-Action')
         if status_action:
             return status_action
-
         status = task.response_status
         assert status is not None
+        location = task.response_headers.get('Location')
+        # handle redirecting to same filepath like source URL
+        if status.startswith('3') and location is not None:
+            redirect_url = url.join(location)
+            if not redirect_url.is_external_to(self.prefix):
+                redirect_path = get_path_from_url(
+                    self.prefix, redirect_url, self.url_to_path
+                )
+                # compare if source path and final path of redirect are same
+                # If they are, apply special logic: consider the target of
+                # the redirection as the page we're supposed to save.
+                same_path = redirect_path == task.path
+                # compare if source url and final url of redirect are different.
+                # If they are not it means a infinite redirect and
+                # it should be handle as other tasks
+                not_same_urls = redirect_url.path != url.path
+
+                if same_path and not_same_urls:
+                    # Only do this if we haven't seen this URL yet.
+                    # If we are, skip this special case. (We're in a redirect
+                    # loop and will probably fail later.)
+                    if redirect_url not in task.urls_redirecting_to_self:
+                        task.urls.add(redirect_url)
+                        task.urls_redirecting_to_self.add(url)
+                        raise RedirectToSamePath()
+
+        if not status_action:
+            # Get a handler for the particular status from configuration
+            status_handler = self.status_handlers.get(status[:3])
 
         # Get a handler for the particular status from configuration
         status_handler = self.status_handlers.get(status[:3])
@@ -540,14 +585,22 @@ class Freezer:
                     generator = import_variable_from_module(generator)
                 self._add_extra_pages(prefix, generator(self.user_app))
             elif isinstance(extra, str):
-                url = urljoin(prefix, decode_input_path(extra))
+                if extra.startswith('/'):
+                    warnings_warn(
+                        f'extra page URL must not start with slash: {extra!r}',
+                        DeprecationWarning,
+                        skip_file_prefixes=(os.path.dirname(__file__),),
+                    )
+                url = prefix.join(decode_input_path(extra))
                 try:
                     self.add_task(
                         url,
                         reason='extra page',
                     )
                 except ExternalURLError:
-                    raise ExternalURLError(f'External URL specified in extra_pages: {url}')
+                    raise ExternalURLError(
+                        f'External URL specified in extra_pages: {url}'
+                    )
             else:
                 generator = extra
                 self._add_extra_pages(prefix, generator(self.user_app))
@@ -562,26 +615,23 @@ class Freezer:
             for path, task in self.inprogress_tasks.items():
                 break
             assert task.asyncio_task is not None
+
             try:
                 await task.asyncio_task
             except Exception as exc:
-                task.update_status(TaskStatus.IN_PROGRESS, TaskStatus.FAILED)
-                self.call_hook('page_failed', hooks.TaskInfo(task))
-                if self.fail_fast:
-                    raise exc
+                task.fail(exc)
             if path in self.inprogress_tasks:
                 raise ValueError(f'{task} is in_progress after it was handled')
 
     @needs_semaphore
     async def handle_one_task(self, task: Task) -> None:
         # Get an URL from the task's set of URLs
-        url_parsed = task.get_a_url()
-        url = url_parsed
+        url = task.get_a_url()
 
         # url_string should not be needed (except for debug messages)
-        url_string = urllib.parse.urlunsplit(url_parsed)
+        url_string = str(url)
 
-        path_info = url_parsed.path
+        path_info = url.path
 
         if path_info.startswith(self.prefix.path):
             path_info = "/" + path_info[len(self.prefix.path):]
@@ -682,6 +732,8 @@ class Freezer:
         except IgnorePage:
             task.update_status(TaskStatus.IN_PROGRESS, TaskStatus.DONE)
             return
+        except RedirectToSamePath:
+            return await self.handle_one_task(task)
 
         await self.saver.save_to_filename(task.path, result_body)
 
@@ -708,7 +760,7 @@ class Freezer:
                         new_links.append(link)
                     links = new_links
                 for link_text in links:
-                    new_url = urljoin(url, link_text)
+                    new_url = url.join(link_text)
                     self.add_task(
                         new_url, external_ok=True,
                         reason=f'linked from: {task.path}',
@@ -723,7 +775,7 @@ class Freezer:
                     link_text, sep, rest = link[1:].partition('>')
                     if not sep:
                         raise ValueError(f'Invalid Link header: {link!r}')
-                    new_url = urljoin(url, link_text)
+                    new_url = url.join(link_text)
                     self.add_task(
                         new_url, external_ok=True,
                         reason=f'Link header from: {task.path}',
@@ -755,8 +807,12 @@ class Freezer:
             if not saved_something:
                 # Get some task (the first one we get by iteration) for the
                 # error message.
+                failing_task = None
                 for task in self.redirecting_tasks.values():
-                    raise InfiniteRedirection(task)
+                    failing_task = task
+                    break
+                if failing_task:
+                    failing_task.fail(InfiniteRedirection(task))
 
     def call_hook(self, hook_name: str, *arguments: Any) -> None:
         for hook in self.hooks.get(hook_name, ()):