+++ freezeyt/types.py
@@ -1,7 +1,6 @@
-from typing import NewType, Mapping, Any, Union, TYPE_CHECKING, Tuple, Iterable
+from typing import Mapping, Any, Union, TYPE_CHECKING, Tuple, Iterable
 from typing import List, TypedDict
 from types import TracebackType
-import urllib.parse
 
 import a2wsgi.asgi_typing
 
@@ -12,10 +11,6 @@ if TYPE_CHECKING:
 Config = Mapping[str, Any]
 
 SaverResult = Union[None, 'DictSaverContents']
-
-# An URL as used internally by Freezeyt.
-# Absolute IRI, with an explicit port if it's `http` or `https`
-AbsoluteURL = NewType('AbsoluteURL', urllib.parse.SplitResult)
 
 ExceptionInfo = Tuple['type[BaseException]', BaseException, TracebackType]